Resolves #324. Add characters classmethod to lorem provider
Sometimes I want to generate a string of characters of a specific length
without any punctuation or capitalization. Adding a characters() method
allows for this. It defaults to a 255-character, mixed case string but
the method can also be called with params that specify length and
whether to exclude uppercase characters.