Added Hungarian phone number provider.
Hungarian phone numbers consist of a two-digit area code (körzethívószám) + a 6 or 7 digit local number (this is being expanded for most cases to 7 digits even in smaller localities). In Hungary, the (0) dial does not exist - out of area dialling is accomplished by calling `06`. For linguistic reasons, phone number grouping is always 3-2-2 or 3-4 (### ## ## or ### ####).