Add title faker
Faker is missing the equivalent to real world titles. A title consists of a
number of words. The first letter of the first word is uppercase and there is
no punctuation.

For example: 'Sapiente sunt omnis'

In applications with a database the title is usually restricted to
certain number of chars. Thus the title faker allows to restrict the
title char number.