ISSUE-1123 handle leap day year shifts
When Faker attempts to generate a date_of_birth on 29th of February on the leap year it fails  if it tries to return a year that is not also a leap year.

This patch fixes the issue by extracting the year adjust to its own function that catches the ValueError and if the starting date is YYYY-02-29 it will adjust the date to YYYY-03-01.