Add a random image provider
Python Image Library is used under the hood to
generate a background and drawing a polygon,
while forwarding the color generation to the
existing color provider.

So as not to force a dependency on Pillow for all
installs, this provider will not work in
absence of this library.

A polygon shape is chosen because its the easiest
drawing method to interface with random coordinates,
as opposed to ellipses or alike that constrain input
parameters, while still providing easily distinguishable
random images, as opposed to random noise bitmaps.