Test for checking the formats defined in providers
For each locale, for each provider, search all the definitions of "formats"
and make sure that all the providers in there (e.g. {{group}}) are valid
and do not emit empty strings. Empty strings are allows only if the group
is not surrounded by spaces. This is a quick way to make sure that no
string is generated with "double spaces", starting spaces or ending spaces.