Make pytuple use _pyiterable instead of pyset
Currently, pytuple uses pyset, which means that if the generates happens
to generate duplicate items, the resulting set will only include on
instances of that item. When asking for explicitly n items, it is
possible to get a number of items that is anywhere between 1 and n inc.

By changing to use a _pyiterable as the pyset and pylist methods do, the
uniqueness of the values generated is no longer required and it behaves
correctly.