Ensures pytuple returns the correct number of items (#1416)
* Add a failing test when pytuple uses pyset

Adds a failing test so that when pytuple asks for 3 ints (where we mock
them to only ever return the number 1), we only get back 1 integer
instead of 3.

* Make pytuple use _pyiterable instead of pyset

Currently, pytuple uses pyset, which means that if the generates happens
to generate duplicate items, the resulting set will only include on
instances of that item. When asking for explicitly n items, it is
possible to get a number of items that is anywhere between 1 and n inc.

By changing to use a _pyiterable as the pyset and pylist methods do, the
uniqueness of the values generated is no longer required and it behaves
correctly.