Separate freely-associated states & US territories (#1828)
* Separate freely-associated states & US territories

There are three sovereign states that are members of the Compact of Free
Association. These states are not US territories, and having
`state_abbr()` include them by default can cause problems with other
tooling.

For the purposes of a valid deliverable address, it's useful to include
these freely-associated states by default. But it's also helpful to be
able to dynamically exclude them. (Furthermore, there's good reason to
simply be correct about how these sovereign states are referred to).

* Fix Black formatting (line length is 120)