Support platform specific second precision
This inspects the platform and defines a platform specific property for
generating random seconds.
For windows this uses `random.randint` as used before and for everything
else `random.uniform`, which returns floats instead of ints.

Then later on when trying to generate a random number of seconds this
proper is then used to get the function that can return values with the
correct behaviour.

I tried setting the function directly on `self` in an `__init__`, but
that worked poorly when seeding the faker after creation, as seeding
recreates the internal random instance, so it requires some lazy
evaluation to be able to point to the correct random instance in all
cases.

Internally the points in time is still just ints. Changing those to
floats is not strictly important, and would just add noise to diff -
making it harder to review.