In the past, Faker would throw an error if the start of a daterange was after the end. This is still the behaviour on Windows but not Unix.
This PR ensures that on both Windows and Unix, an exception is
thrown. This is "backwards compatible" to older versions of Faker
and compatible across platforms.

On all platforms, the exception is a ValueError that contains
the string "empty range" and this can be used by callers
to detect the problem.