Overload hash generation functions to make return types more specific (#2047)
Before this change, the return type for these functions was always
`bytes| str`. However, the type can be narrowed based on the value
passed to `raw_output`.

This change adds overload annotations for the two allowable values to
allow type checkers to narrow the type and correctly infer the exact
return type.

Resolves #2046

Co-authored-by: Flavio Curella <89607+fcurella@users.noreply.github.com>